LNC_LBP.py constructs the Lifted Graph for the Smoking example.
Note that this version only builds the lifted graph, and the superEdge weights and liftedBP are not taken into account yet.
Also the LNC process investigates every single grounding individually which probably is inefficient. I should consider all the data that are absent from all evidences as one item duing the projection (in parvar_grnd).