LNC_LBP.py constructs the Lifted Graph for the Smoking example.
[1] In this version, the lifted graph is built using the superNodes and superFeatures. Also the liftedBP is implemented by modifying node.py and graph.py codes.
Note that for building this graph, I have omitted the factors that connect only two superNodes, though I am not sure it is the right thing to do. For example, this approach completely ignores Friends(B,C) and Friends(C,B) when they are seen as evidence and put in one supernode.
[2] Also the LNC process investigates every single grounding individually which probably is inefficient. I should consider all the data that are absent from all evidences as one item duing the projection (in parvar_grnd).