Lifted Network Construction (LNC) and Lifted Beleif Propagation (LiftedBP)
Nothing that much changed. I have just realized that if e.g. for the case 'names' = {A, B, C}, Smoke(A) is evidence and B & C are unknown, the Lifted BP results in incorrect probabilities, since it completely drops B-C internal structure from the graph. But if I use the marginal results of B-C supernode as factor potentials for a new graph that is made only for the B-C structure, it led to the correct results. (some code added to the end of simple5.py). But similar approach did not work e.g. for {A, B, C, D, F} where evidences are Smoke(A) and Friends(C,D).